introduce mirror URL-s

+++ fetch_packages.py
@@ -7,20 +7,22 @@ import errno
 import re
 import shutil
 import subprocess
-import sys, getopt
+import sys
+import getopt
 import platform
 from time import sleep
 
 _RETRIES = 5
 _OPT_VERBOSE = None
 _OPT_DRY_RUN = None
-_PACKAGE_CACHE='/tmp/cache/' + os.environ['USER'] + '/webui_third_party'
-_NODE_MODULES='./node_modules'
-_TMP_NODE_MODULES=_PACKAGE_CACHE + '/' + _NODE_MODULES
+_PACKAGE_CACHE = '/tmp/cache/' + os.environ['USER'] + '/webui_third_party'
+_NODE_MODULES = './node_modules'
+_TMP_NODE_MODULES = _PACKAGE_CACHE + '/' + _NODE_MODULES
 _TAR_COMMAND = ['tar', '--no-same-owner']
 _CACHED_PKG_DISTROS = ('Ubuntu', 'Red Hat', 'CentOS', 'darwin')
 
 from lxml import objectify
+
 
 def getFilename(pkg, url):
     element = pkg.find("local-filename")
@@ -33,13 +35,15 @@ def getFilename(pkg, url):
         filename = m.group(1)
     return filename
 
+
 def setTarCommand():
     if isTarGnuVersion():
-        print 'GNU tar found. we will skip the no-unknown-keyword warning'
+        print('GNU tar found. we will skip the no-unknown-keyword warning')
         global _TAR_COMMAND
         _TAR_COMMAND = ['tar', '--no-same-owner', '--warning=no-unknown-keyword']
     else:
-        print 'No GNU tar. will use default tar utility'
+        print('No GNU tar. will use default tar utility')
+
 
 def isTarGnuVersion():
     cmd = subprocess.Popen(['tar', '--version'],
@@ -50,13 +54,15 @@ def isTarGnuVersion():
         return True
     return False
 
+
 def getTarDestination(tgzfile, compress_flag):
-    cmd = subprocess.Popen( _TAR_COMMAND + [ '--exclude=.*','-' + compress_flag + 'tf', tgzfile],
+    cmd = subprocess.Popen(_TAR_COMMAND + ['--exclude=.*', '-' + compress_flag + 'tf', tgzfile],
                            stdout=subprocess.PIPE)
     (output, _) = cmd.communicate()
     (first, _) = output.split('\n', 1)
     fields = first.split('/')
     return fields[0]
+
 
 def getZipDestination(tgzfile):
     cmd = subprocess.Popen(['unzip', '-t', tgzfile],
@@ -64,17 +70,19 @@ def getZipDestination(tgzfile):
     (output, _) = cmd.communicate()
     lines = output.split('\n')
     for line in lines:
-        print line
+        print(line)
         m = re.search(r'testing:\s+([\w\-\.]+)\/', line)
         if m:
             return m.group(1)
     return None
 
+
 def getFileDestination(file):
     start = file.rfind('/')
     if start < 0:
         return None
-    return file[start+1:]
+    return file[start + 1:]
+
 
 def ApplyPatches(pkg):
     stree = pkg.find('patches')
@@ -89,11 +97,9 @@ def ApplyPatches(pkg):
             print "Patching %s <%s..." % (' '.join(cmd), str(patch))
         if not _OPT_DRY_RUN:
             fp = open(str(patch), 'r')
-            proc = subprocess.Popen(cmd, stdin = fp)
+            proc = subprocess.Popen(cmd, stdin=fp)
             proc.communicate()
 
-#def VarSubst(cmdstr, filename):
-#    return re.sub(r'\${filename}', filename, cmdstr)
 
 def GetOSDistro():
     distro = ''
@@ -102,9 +108,10 @@ def GetOSDistro():
     else:
         try:
             return platform.linux_distribution()[0]
-        except:
+        except Exception:
             pass
     return distro
+
 
 def DownloadPackage(url, ccfile, pkg):
     md5 = pkg.md5
@@ -128,7 +135,7 @@ def DownloadPackage(url, ccfile, pkg):
                     distro = cached_pkg
                     found = True
                     break
-            if found == False:
+            if not found:
                 # Remove the $distro from the url and try
                 url = url.replace('/$distro', '')
                 # Change the pkg format to npm download the dependencies
@@ -153,14 +160,14 @@ def DownloadPackage(url, ccfile, pkg):
                     pass
         print url
 
-    #Check if the package already exists
+    # Check if the package already exists
     if os.path.isfile(ccfile):
         md5sum = FindMd5sum(ccfile)
         if md5sum == md5:
             return pkg
         else:
             os.remove(ccfile)
-    
+
     retry_count = 0
     while True:
         subprocess.call(['wget', '--no-check-certificate', '-O', ccfile, url])
@@ -179,6 +186,7 @@ def DownloadPackage(url, ccfile, pkg):
             raise RuntimeError("MD5sum %s, expected(%s) dosen't match for the "
                                "downloaded package %s" % (md5sum, md5, ccfile))
     return pkg
+
 
 def ProcessPackage(pkg):
     print "Processing %s ..." % (pkg['name'])
@@ -199,7 +207,7 @@ def ProcessPackage(pkg):
             else:
                 print 'mkdirs of ' + _NODE_MODULES + ' failed.. Exiting..'
                 return
-        #ccfile = _NODE_MODULES + '/' + filename
+        # ccfile = _NODE_MODULES + '/' + filename
     pkg = DownloadPackage(url, ccfile, pkg)
 
     #
@@ -232,9 +240,8 @@ def ProcessPackage(pkg):
     rename = pkg.find('rename')
     if rename and pkg.format == 'npm-cached':
         rename = _NODE_MODULES + '/' + str(rename)
-    if rename and os.path.isdir(str(rename)):
-        if not _OPT_DRY_RUN:
-          shutil.rmtree(str(rename))
+    if rename and os.path.isdir(str(rename)) and not _OPT_DRY_RUN:
+        shutil.rmtree(str(rename))
 
     elif dest and os.path.isdir(dest):
         if _OPT_VERBOSE:
@@ -292,25 +299,26 @@ def ProcessPackage(pkg):
             if os.path.exists(newDir + '/' + pkg['name']):
                 cmd = npmCmd
             else:
-		try:
-                   p = subprocess.Popen(cmd, cwd = cd)
-                   ret = p.wait()
-                   if ret is not 0:
-                       sys.exit('Terminating: ProcessPackage with return code: %d' % ret);
-                   cmd = npmCmd
-		except OSError:
-		   print ' '.join(cmd) + ' could not be executed, bailing out!'
-		   return
+                try:
+                    p = subprocess.Popen(cmd, cwd=cd)
+                    ret = p.wait()
+                    if ret is not 0:
+                        sys.exit('Terminating: ProcessPackage with return code: %d' % ret)
+                    cmd = npmCmd
+                except OSError:
+                    print(' '.join(cmd) + ' could not be executed, bailing out!')
+                    return
 
-        p = subprocess.Popen(cmd, cwd = cd)
+        p = subprocess.Popen(cmd, cwd=cd)
         ret = p.wait()
         if ret is not 0:
-            sys.exit('Terminating: ProcessPackage with return code: %d' % ret);
+            sys.exit('Terminating: ProcessPackage with return code: %d' % ret)
 
     if rename and dest:
         os.rename(dest, str(rename))
 
     ApplyPatches(pkg)
+
 
 def FindMd5sum(anyfile):
     if sys.platform == 'darwin':
@@ -323,23 +331,25 @@ def FindMd5sum(anyfile):
     md5sum = stdout.split()[0]
     return md5sum
 
+
 def main(filename):
     tree = objectify.parse(filename)
     root = tree.getroot()
-    #Check which version of tar is used and skip warning messages.
+    # Check which version of tar is used and skip warning messages.
     setTarCommand()
     for object in root.iterchildren():
         if object.tag == 'package':
             ProcessPackage(object)
 
+
 if __name__ == '__main__':
     try:
-        opts,args = getopt.getopt(sys.argv[1:],"f:",["file="])
+        opts, args = getopt.getopt(sys.argv[1:], "f:", ["file="])
     except getopt.GetoptError:
         raise RuntimeError("Error in parsing the options/arguments")
     xmlfile = None
-    for opt,arg in opts:
-        if opt in ("-f","--file"):
+    for opt, arg in opts:
+        if opt in ("-f", "--file"):
             xmlfile = arg
     os.chdir(os.path.dirname(os.path.realpath(__file__)))
     try:
@@ -347,7 +357,7 @@ if __name__ == '__main__':
     except OSError:
         pass
 
-    if xmlfile == None:
+    if xmlfile is None:
         main('packages.xml')
     else:
         main(xmlfile)
